+++ documentation/docs/release-notes.md
@@ -9,6 +9,7 @@ title: Release notes&#58;
 - Allow to force the reloading of the SAML metadata
 - Reinforce security by checking OIDC logout requests (can be disabled via `OidConfiguration.setLogoutValidation(false)`)
 - Retrieving OIDC resources such as keys from a remote IDP now recognizes the OIDC configuration for remote hostname verification.
+- OAuth2 credentials can now be serialized from/to JSON correctly using an intermediate object to carry the access token.
 
 **v6.0.4**:
 - OIDC support: set the profile identifier from the subject of the userinfo endpoint if need be